[Refactor]: attendanceId 타입 수정

+++ src/main/java/live/smoothing/user/attendance/entity/Attendance.java
@@ -16,7 +16,7 @@ public class Attendance {
     @Id
     @Column(name = "attendance_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int attendanceId;
+    private Long attendanceId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
